Mention that ncurses 6.4 addresses #1027435
The upstream change fixes the problem as far as ncurses is concerned,
but with vim versions before 9.0.1117 users will still experience it,
so I prefer to keep the bug report open.

Once a fixed vim arrives in testing we can add versioned Breaks to
ncurses-{base,term} to avoid breakage on partial upgrades.