Run autoconf-dickey rather than autoreconf-dickey in debian/rules
Unfortunately autoreconf-dickey picks up and processes the quilt
backup file of configure.in below .pc/ directory.  As the other
expected files are not present there, this fails and randomly leads to
FTBFS, depending on the order in which autoreconf-dickey processes the
directories.

As pointed out by Thomas Dickey, we do not need to run the whole
autotools machinery and can make do with autoconf-dickey which only
runs in one directory at a time.  Since we should regenerate both
./configure and test/configure and dh_autoreconf may only be run once,
this needs to be done by a script, as suggested by the dh_autoreconf
manpage.

Closes: #1035621