Set ARFLAGS = -crv to ensure deterministic static libraries
We should build the static libraries reproducibly even if they are not
stripped, e.g. when dh_strip fails (see #875780 and #1029404), or when
"nostrip" is in DEB_BUILD_OPTIONS.

Upstream sets ARFLAGS=-curvU by default, which is good during
development as it avoids unnecessary rebuilds of the static libraries,
but bad if you care about reproducibility.

Note that ARFLAGS = -crvD would be a bit clearer, but unfortunately it
does not work.  The CF_AR_FLAGS macro in aclocal.m4 compares ARFLAGS
against a set of known values, and if it is not found, the build
fails.