Stop stripping libncurses++w.a manually
This effectively reverts commit af4606ad8f0a.

Commit bd2267f6ffb7 ("Set ARFLAGS = -crv to ensure deterministic
static libraries") already fixed the reproducibility problem for good,
and a test build on i386 no longer triggered the dh_strip bug #875780,
presumably due to recent changes in gcc or binutils.

Of course there is no guarantee that the problem of unstripped static
libraries will not show up on other architectures or on i386 in the
future, but at least the package is still reproducible, and a few
extra hundred kilobytes in a -dev package are not such a big deal.