Move the cons25-debian terminfo entry to ncurses-term
It had only been used in Debian GNU/kFreeBSD 6.0 as an alternative to
the standard cons25 entry for the console, see the long and painful
discussion in https://bugs.debian.org/607662.

Debian 7 switched to TERM=xterm for the console, as did FreeBSD 9.
So there should not really be any users of cons25-debian left.