Tighten the dependency of ncurses-term on ncurses-base
Now that the terminfo database is no longer spread over two
directories, we need to be extra careful when moving entries from
ncurses-term to ncurses-base or vice versa.  In particular, care has
to be taken not to lose some entry on partial updates.

Currently this means we would have to bump the "Depends" of
ncurses-term on ncurses-base as well as the "Breaks" of ncurses-base
on ncurses-term, which is error-prone.  By ensuring that both packages
are updated at the same time, only "Replaces" relationships need to be
updated.

In the future we could also Drop the "Breaks" from ncurses-base, once
the strict dependencies are in Debian stable.