Remove xterm+focus from alacritty terminfo entry
Vim versions before 9.0.1619 misbehave with TERM=alacritty when
switching focus to or from the terminal window.  While this has
already been fixed in unstable, vim currently FTBFS on s390x and so
cannot migrate to testing.  The upstream bug at
https://github.com/vim/vim/issues/12586 is still under discussion.

Once a fixed vim is in testing we should drop this patch and consider
bumping the Breaks against vim-common in ncurses-term.