Attempt to fix intermittent dh_missing failure
Occasionally a full build causes an error in dh_missing:

dh_missing -i --fail-missing
dh_missing: warning: usr/lib32/libncurses.so exists in debian/tmp but is not installed to anywhere (related file: "debian/tmp/usr/lib/x86_64-linux-gnu/libncurses.so")
dh_missing: warning: usr/lib32/libncursesw.so exists in debian/tmp but is not installed to anywhere (related file: "debian/tmp/usr/lib/x86_64-linux-gnu/libncursesw.so")
dh_missing: error: missing files, aborting

Which is rather weird, because the files have been created in the
install-arch target of debian/rules.  This had happened a few times in
my local test builds, but not elsewhere as far as I am aware.
However, the reprotest job in the Salsa CI pipeline triggered it now,
and retrying the job did not help.

https://salsa.debian.org/debian/ncurses/-/jobs/4377948/raw

I suspect there is a bug in dh_missing.  The closest report in the
Debian BTS seems to be https://bugs.debian.org/993683, where the
debhelper maintainer gave the advice to prefix the path to the
offending files with "debian/tmp".

Our situation is a bit different because there should be no top level
usr/lib32 directory, and I am also not able to reliably reproduce the
problem locally.  Still, adding the "debian/tmp" prefix cannot hurt
and is also a bit more consistent, as the other source files in
lib{32,64}ncurses-dev.install are not relative to debian/tmp.