Skip building static libraries along the shared ones
These are not used and replaced by the ones from the dedicated static
builds, so we do not have to create them in the first place.

As the manual installation of the static libraries from the build
directories fails to create the libncurses.a -> libcurses.a symlink,
fix that up via libncurses-dev.links.

Test builds showed that the resulting binary packages are identical.