Reorder the dh_* commands in the binary-{arch,indep} targets
We ran them in a different from order than dh(1) would do, for
instance the dh_installdirs calls (which are currently no-ops as there
is no .dirs file in debian/) were oddly placed.

Package contents are unchanged, but the Installed-Size is increased by
1K in ncurses-base and the library packages due to swapping of the
dh_gencontrol and dh_installdeb commands.  The latter installs the
"conffiles" and "triggers" files into the DEBIAN directories, and
those are taken into account by dpkg-gencontrol when it computes the
Installed-Size field.