Add a versioned Provides on foot-terminfo to ncurses-term
The foot package has a dependency on foot-terminfo (>= 1.5.3-4) which
would no longer be fulfilled once they upgrade ncurses-term.  This
could be bad if ncurses migrates to testing before foot, as foot and
ncurses-term cannot both be installed.  Avoid such a situation by
providing foot-terminfo in ncurses-term.

Partial upgrades from stable will also be easier, since upgrading
ncurses-term only forces the removal (or upgrade) of the foot-terminfo
package and not of the foot package.