Fix cross-building for arm64 and amd64
We cannot use the same flags for the host and build compilers, because
the output of dpkg-buildflags is architecture dependent.  In version
1.22.0 dpkg-buildflags started to add -mbranch-protection=standard on
arm64 and -fcf-protection on amd64, those flags are usually not
supported by the native compiler.

The dpkg-buildflags manpage does not currently mention any of this,
nor does it give a hint how to override the architecture.  Some ideas
have been mentioned in https://bugs.debian.org/1021292, but apparently
nothing has come forth yet.