Install upstream libtinfo.so symlink
We could not do this hitherto, because the libtinfo symlink was in a
different directory than the target, but now there is no reason to
treat it differently from the other .so symlinks.

The only difference is that libtinfo.so now points to libtinfo.so.6.4
rather than libtinfo.so.6, but that does not matter.