Ensure that ncurses6-config.1 gets installed into debian/tmp
Up to now we installed it directly from the build tree which has the
disadvantage that edit_man is not run on it to fix the reference from
curses(3X) to ncurses(3ncurses).  Copy it from the non-wide build tree
to the wild one instead, where "make install" picks it up.

This relies on man/Makefile to process all *-config.1 manpages, which
it has been doing since the 20100313 patchlevel.

Closes: #1060211