Add a build profile to check terminfo database
This will mostly be useful when bumping the Debian revision only,
e.g. to test if there are unexpected changes to the terminfo files
after applying patches for the next tic CVE.

It is also helpful to compare the new database against the current one
for the ongoing upstream effort to trim the large terminfo.src file
with use=…  clauses.  There have been cases recently where the binary
terminfo files contained unintended changes.

https://lists.gnu.org/archive/html/bug-ncurses/2024-01/msg00003.html
https://lists.gnu.org/archive/html/bug-ncurses/2024-01/msg00074.html

For now we use the system's infocmp version for the report.  Running
the just built one could be a possible improvement, but does not work
in cross builds.  Use diff's --no-dereference option so that entries
with multiple aliases are only reported once.