Fix manpage-references test architecture restriction
It is possible that the libncurses-dev package we are testing is not
of the same architecture as dpkg, and then the test would fail
unnecessary if the native libncurses-dev package is not installed.
This seems to be the case for the i386 architecture on Ubuntu's
autopkgtest infrastructure, where the test failed anyway for a
different reason.

https://bugs.launchpad.net/ubuntu/+source/ncurses/+bug/2052951

Instead we should figure out which versions of libncurses-dev are
actually installed and test all of them.