Add another smoke autopkgtest for backspace/delete keys
Debian Policy demands that kbs=^? is the preferred behavior of the
backspace key if there is a choice.  Test several terminfo entries for
that.  The test could fail if the Debian patches had not applied
during build, or if they had been rebased incorrectly, for instance.

Chapter 9.8 (Keyboard configuration) of the Debian Policy mentions
that kdch1=\E[3~ is obligatory for several values of TERM.