Stop patching terminfo.src for Debian-specific terminfo aliases
For over a decade the Debian-specific terminfo entries xterm-debian
and rxvt-m have been aliases for xterm and rxvt-basic, respectively.
The recent upstream changes to misc/terminfo.src required updating our
patches, which is tedious.

Instead of patching the source, create the two symlinks by hand.  We
do that in debian/rules rather than via ncurses-base.links, so that
they are not reported as missing in the pkg.ncurses.terminfocheck
build profile.  The only difference in the resulting binary terminfo
files is that the alias is no longer a known name for the terminal,
but the aliases should be deprecated anyway.