Update patch 01-debian-no-ada-doc.diff
In version 6.3.20211021-5 the libncursesada package dropped the -doc
binary package and put the documentation into the libncursesadaN-dev
package, invalidating the link in our patch.  Version 6.3.20211021-10
of libncursesada switched to an unversioned -dev package, so the
location of the documentation should be stable again and not change
with every gnat transition.